requirements.txt
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
numpy
Pillow
torch
tqdm
transformers
huggingface-hub
sentence-transformers
open_clip_torch
nltk
